+++ authentication/lib/refinery/authenticated_system.rb
@@ -15,7 +15,7 @@ module Refinery
     end
 
     # This just defines the devise method for after sign in to support
-    # engine namespace isolation...
+    # extension namespace isolation...
     def after_sign_in_path_for(resource_or_scope)
       scope = Devise::Mapping.find_scope!(resource_or_scope)
       home_path = "#{scope}_root_path"