+++ core/lib/refinery/errors.rb
@@ -6,7 +6,7 @@ module Refinery
   class RefineryError < StandardError
   end
 
-  # Raised when an engine has not been properly defined. See the exception message for further
+  # Raised when an extension has not been properly defined. See the exception message for further
   # details
   class InvalidEngineError < RefineryError
   end