+++ core/spec/lib/generators/refinery/engine/engine_generator_spec.rb
@@ -13,7 +13,7 @@ module Refinery
     end
 
     specify do
-      destination_root.should have_structure do
+      destination_root.should have_structure {
         directory "vendor" do
           directory "extensions" do
             directory "rspec_product_tests" do
@@ -78,7 +78,7 @@ module Refinery
             end
           end
         end
-      end
+      }
     end
   end
 end