+++ pages/lib/generators/refinery/pages/pages_generator.rb
@@ -11,7 +11,7 @@ module Refinery
       append_file 'db/seeds.rb', :verbose => true do
         <<-EOH
 
-# Added by Refinery CMS Pages engine
+# Added by Refinery CMS Pages extension
 Refinery::Pages::Engine.load_seed
         EOH
       end