+++ testing/lib/refinery/testing.rb
@@ -11,11 +11,11 @@ module Refinery
         @root ||= Pathname.new(File.expand_path('../../../', __FILE__))
       end
 
-      # Load the factories of all currently loaded engines
+      # Load the factories of all currently loaded extensions
       def load_factories
-        Refinery.engines.each do |engine_const|
-          if engine_const.respond_to?(:factory_paths)
-            engine_const.send(:factory_paths).each do |path|
+        Refinery.extensions.each do |extension_const|
+          if extension_const.respond_to?(:factory_paths)
+            extension_const.send(:factory_paths).each do |path|
               FactoryGirl.definition_file_paths << path
             end
           end